refactor : query 패키지 생성으로 임포트 변경
