+++ src/main/java/com/fourthread/ozang/module/domain/clothes/repository/ClothesAttributeDefinitionRepository.java
@@ -1,6 +1,7 @@
 package com.fourthread.ozang.module.domain.clothes.repository;
 
 import com.fourthread.ozang.module.domain.clothes.entity.ClothesAttributeDefinition;
+import com.fourthread.ozang.module.domain.clothes.repository.query.ClothesAttributeDefinitionRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
